+++ services/rds/struct_latency_top_n_item.go
@@ -17,7 +17,7 @@ package rds
 
 // LatencyTopNItem is a nested struct in rds response
 type LatencyTopNItem struct {
-	SQLExecuteTimes int    `json:"SQLExecuteTimes" xml:"SQLExecuteTimes"`
-	AvgLatency      int    `json:"AvgLatency" xml:"AvgLatency"`
+	SQLExecuteTimes int64  `json:"SQLExecuteTimes" xml:"SQLExecuteTimes"`
+	AvgLatency      int64  `json:"AvgLatency" xml:"AvgLatency"`
 	SQLText         string `json:"SQLText" xml:"SQLText"`
 }
