+++ services/rds/struct_qps_top_n_item.go
@@ -17,6 +17,6 @@ package rds
 
 // QPSTopNItem is a nested struct in rds response
 type QPSTopNItem struct {
-	SQLExecuteTimes int    `json:"SQLExecuteTimes" xml:"SQLExecuteTimes"`
+	SQLExecuteTimes int64  `json:"SQLExecuteTimes" xml:"SQLExecuteTimes"`
 	SQLText         string `json:"SQLText" xml:"SQLText"`
 }
